donate update

+++ database/migrations/2020_07_25_080611_create_table_donate.php
@@ -19,8 +19,8 @@ class CreateTableDonate extends Migration
             $table->string("donate_image")->nullable();
             $table->text("donate_desc");
             $table->text("donate_content");
-            $table->decimal("raisermoney",12,4)->default(0);
-            $table->decimal("goalmoney",12,4);
+            $table->bigInteger("raisermoney")->default(0);
+            $table->bigInteger("goalmoney");
             $table->unsignedBigInteger("user_id");
             $table->foreign("user_id")->references("id")->on("users");
             $table->timestamps();
